requirements.txt
annotated-types==0.7.0
anyio==4.10.0
audioread==3.0.1
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
decorator==5.2.1
fastapi==0.116.1
h11==0.16.0
idna==3.10
Jinja2==3.1.6
joblib==1.5.1
lazy_loader==0.4
librosa==0.11.0
llvmlite==0.44.0
MarkupSafe==3.0.2
msgpack==1.1.1
numba==0.61.2
numpy==2.2.6
packaging==25.0
platformdirs==4.3.8
pooch==1.8.2
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
python-multipart==0.0.20
requests==2.32.4
scikit-learn==1.7.1
scipy==1.16.1
sniffio==1.3.1
soundfile==0.13.1
soxr==0.5.0.post1
starlette==0.47.2
threadpoolctl==3.6.0
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
uvicorn==0.35.0

# Audio Generation (needs >= 1.9.0 for text_to_sound_effects)
elevenlabs>=1.9.0

# Environment Variables
python-dotenv==1.0.0

# Additional dependencies
pydantic-settings==2.1.0
